requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.2
frozenlist==1.2.0
future==0.18.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
matplotlib==3.5.0
multidict==5.2.0
numpy==1.21.4
opencv-python-headless==4.5.3.56
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
